+++ pkg/common/envoy/request/metadata_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package envoy
+package request
 
 import (
 	"testing"
